Add files via upload
This website implements express.js and catches all possible routes. The website already has 4 existing players displayed on the homepage. If the user wanted to find out more information about any of the players, he or she must click on the player's image and they will be sent to another page. If the user wanted the website to display a new player that was not on the homepage, they can do that by filling out the form and uploading an image of that player. The information from the form would be sent to the db.json database file and the new player would also be displayed on the homepage and functions similarly to the players that existed prior.